requirements.txt
gym-trading-env
yfinance
matplotlib
numpy
pandas
pytest
torch
pytorch-lightning
absl-py
